Move GetAssistancePage out of workspace folder to main pages folder
