+++ src/pages/GetAssistancePage.js
@@ -1,18 +1,18 @@
 import React from 'react';
 import {View, ScrollView} from 'react-native';
 import PropTypes from 'prop-types';
-import ScreenWrapper from '../../components/ScreenWrapper';
-import withLocalize, {withLocalizePropTypes} from '../../components/withLocalize';
-import KeyboardAvoidingView from '../../components/KeyboardAvoidingView';
-import HeaderWithCloseButton from '../../components/HeaderWithCloseButton';
-import MenuItem from '../../components/MenuItem';
-import Navigation from '../../libs/Navigation/Navigation';
-import styles from '../../styles/styles';
-import Text from '../../components/Text';
-import RequestCallIcon from '../../../assets/images/request-call.svg';
-import * as Expensicons from '../../components/Icon/Expensicons';
-import * as Report from '../../libs/actions/Report';
-import ROUTES from '../../ROUTES';
+import ScreenWrapper from '../components/ScreenWrapper';
+import withLocalize, {withLocalizePropTypes} from '../components/withLocalize';
+import KeyboardAvoidingView from '../components/KeyboardAvoidingView';
+import HeaderWithCloseButton from '../components/HeaderWithCloseButton';
+import MenuItem from '../components/MenuItem';
+import Navigation from '../libs/Navigation/Navigation';
+import styles from '../styles/styles';
+import Text from '../components/Text';
+import RequestCallIcon from '../../assets/images/request-call.svg';
+import * as Expensicons from '../components/Icon/Expensicons';
+import * as Report from '../libs/actions/Report';
+import ROUTES from '../ROUTES';
 
 const propTypes = {
     /** Route object from navigation */
